+++ api/v1/routes/projects/uploadProject.js
@@ -31,7 +31,7 @@ module.exports = (app, utils) => {
         }
 
         // make sure its been 8 minutes since last upload
-        if (await utils.UserManager.getLastUpload(username) > Date.now() - utils.uploadCooldown) {
+        if (await utils.UserManager.getLastUpload(username) > Date.now() - utils.uploadCooldown && (!await utils.UserManager.isAdmin(username) && !await utils.UserManager.isModerator(username))) {
             await unlink();
             return utils.error(res, 400, "Uploaded in the last 8 minutes");
         }
